==> notebooks/0_feature_extraction.ipynb <==
import os
from collections import defaultdict
import string
import re
import numpy as np
from tqdm import tqdm
from tqdm._tqdm_notebook import tqdm_notebook
import pandas as pd
from multiprocessing import Pool
tqdm_notebook.pandas()

def mkdir(d):
    if not os.path.exists(d):
        os.makedirs(d)

## Data Loading

We first load the Yelp dataset as well as the annotations from [1].

[1] Hernández-Rubio, M., Cantador, I., Bellogín, A. 2019. A comparative analysis of recommender systems based on item aspect opinions extracted from user reviews. User Modeling and User-Adapted Interaction 29(2), 381-441.

# Dataframe for all reviews.
df_review = pd.read_csv("../data/yelp_data/review.csv").drop(["useful", "funny", "cool"], axis=1)
df_review = df_review.rename(columns={"business_id": "item_id"})

# Dataframe for all restaurants.
df_item = pd.read_csv("../data/yelp_data/business.csv")
df_item = df_item.rename(columns={"business_id": "item_id"})
df_item["is_restaurant"] = df_item[["categories"]].apply(lambda x: "Restaurants" in x["categories"], axis=1)
df_item = df_item[df_item["is_restaurant"]]
df_item = df_item[["item_id", "name", "city", "state", "postal_code", "latitude", "longitude", "stars"]]

# Dataframe for words representing aspects.
df_aspects = pd.read_csv("../data/aspect/aspects_restaurants.csv", header=0, names=["aspect", "word"])
aspect_dict = defaultdict(list)
aspect_reverse_dict = {}
for i, row in df_aspects.iterrows():
    aspect_dict[row["aspect"]].append(row["word"])
    aspect_reverse_dict[row["word"]] = row["aspect"]

# Dataframe for words representing aspect sentiment orientations.
df_so = pd.read_csv("../data/aspect/lexicon_restaurants.csv", header=0, names=["aspect", "word", "score"])
so_dict = defaultdict(dict)
so_reverse_dict = {}
for i, row in df_so.iterrows():
    so_dict[row["aspect"]][row["word"]] = row["score"]
    so_reverse_dict[row["word"]] = row["aspect"]

# preview the review dataframe
df_review.iloc[:5]

## Item Profiling

We propose to use latent factor-based item profiles to model restaurants in the dataset. Concretely, for each item $i$ and aspect $k$ extracted in [1], we compute two values for the (item, aspect) pair by going through all reviews for $i$:

- $\psi_{i,k}$: the proportion of all review contents discussing about aspect $k$;
- ${so}_{i,k}$: the average sentiment orientation of all words in aspect $k$.

We define the item aspect weight for item $i$ and aspect $k$ to be $$w_{i,k} = \psi_{i,k} \cdot {so}_{i,k}.$$

def text_aspect_counts(text):
    ''' Given a block of text, count the number of words in each aspect. 
    
        Returns: a list of counts with indices corresponding to so_dict.
    '''
    words = re.sub('['+string.punctuation+']', '', text.lower()).split()
    counts = [0] * len(so_dict)
    for word in words:
        if word in aspect_reverse_dict:
            aspect_name = aspect_reverse_dict[word]
            aspect_ix = list(so_dict.keys()).index(aspect_name)
            counts[aspect_ix] += 1
    return counts

def text_so_measure(text):
    ''' Given a block of text, calculate the sum of sentiment orientation 
        values as well as their counts.
    
        Returns:
            sums: the sum of so values
            counts: count of values used for the sums
    '''
    words = re.sub('['+string.punctuation+']', '', text.lower()).split()
    sums = [0] * len(so_dict)
    counts = [0] * len(so_dict)
    for word in words:
        if word in so_reverse_dict:
            aspect_name = so_reverse_dict[word]
            aspect_ix = list(so_dict.keys()).index(aspect_name)
            word_so = so_dict[aspect_name][word]
            sums[aspect_ix] += word_so
            counts[aspect_ix] += 1
    return sums, counts

def item_aspect_weights(item_id):
    ''' Calculates w_{i,k} wrt. all k for item i. '''
    reviews = df_review[df_review['item_id'] == item_id]["text"].tolist()
    
    # calculating $\psi$
    psi = np.zeros((len(so_dict),))
    for review in reviews:
        psi += text_aspect_counts(review)
    psi = psi.astype(float) / np.sum(psi)
    
    # calculating so
    sums = np.zeros((len(so_dict),))
    counts = np.zeros((len(so_dict),))
    for review in reviews:
        review_sum, review_count = text_so_measure(review)
        sums += review_sum
        counts += review_count
    counts[counts == 0] = 1
    so = sums / counts
    
    return psi * so

# example usage
item_aspect_weights('5ubokMNw8qfbX2WtxgJG1Q')

# add column for item profile
df_item['aspect_weights'] = 0
df_item['aspect_weights'] = df_item['aspect_weights'].astype(object)

for i, row in tqdm_notebook(df_item.iterrows(), total=len(df_item)):
    df_item.at[i, 'aspect_weights'] = item_aspect_weights(row['item_id'])

# save produced item profiles
mkdir('../data/generated')
df_item.to_json('../data/generated/item.json')

## User Profiling

We propose to use implicit aspect-based user profiles to model users. Concretely, we compute a vector of length $K$ for each user, where $K$ is the number of extracted aspects in [1]. Each item in the vector, which we denote as $w_{u,k}$, can be computed as $$w_{u,k} = \sum_{\{(u,i,r), r \neq \emptyset\}} r(u,i) \cdot w_{i,k}.$$

def user_aspect_weights(user_id):
    reviews = df_review[df_review["user_id"] == user_id][["item_id", "stars"]]
    w_uk = np.zeros((len(so_dict),))
    for i, row in reviews.iterrows():
        item_row = df_item.loc[df_item["item_id"] == row["item_id"]]
        if len(item_row) == 0:
            continue
        w_uk += float(row["stars"]) * item_row.iloc[0]["aspect_weights"]
    return w_uk

# example usage
user_aspect_weights('bv2nCi5Qv5vroFiqKGopiw')

# add column for user profile
df_user['aspect_weights'] = 0
df_user['aspect_weights'] = df_user['aspect_weights'].astype(object)

for i, row in tqdm_notebook(df_user.iterrows(), total=len(df_user)):
    df_user.at[i, 'aspect_weights'] = user_aspect_weights(row['user_id'])

# save produced user profiles
mkdir('../data/generated')
df_user.to_json('../data/generated/user.json')
