==> notebooks/1_recommendation_engine.ipynb <==
import os
import numpy as np
import pandas as pd
import faiss
from tqdm import tqdm

## Data Loading

Load yelp dataset along with pre-computed user and item vectors.

# read datasets
df_item = pd.read_json('../data/generated/item.json')
df_user = pd.read_json('../data/generated/user.json')

df_review = pd.read_csv('../data/yelp_data/review.csv')

df_item.head(n=1)

df_user.head(n=1)

df_review.head(n=1)

## Preparations

Load index for fast nearby user retrieval and function for converting user information to feature vectors.

# initialize the user index
d = len(df_user.iloc[0]['aspect_weights'])
index = faiss.IndexFlatIP(d)

# populate the index with user vectors
user_vectors = np.array(df_user['aspect_weights'].to_list()).astype(np.float32)
user_vectors = user_vectors / np.linalg.norm(user_vectors, axis=1, keepdims=True)
index.add(user_vectors)

def get_user_vector(reviews):
    # sample input: [{"item_id": "xyz", "rating": 5}]
    # output: user vector
    w_uk = np.zeros((d,))
    for review in reviews:
        item_row = df_item.loc[df_item["item_id"] == review["item_id"]]
        if len(item_row) == 0:
            continue
        w_uk += float(review["rating"]) * np.array(item_row.iloc[0]["aspect_weights"]).astype(float)
    return w_uk / np.linalg.norm(w_uk)

user_vector = get_user_vector([{"item_id": "XOSRcvtaKc_Q5H1SAzN20A", "rating": 5}])

## Item Rating Prediction

Inspired by section 4.2.2 of [1], we design our method for predicting ratings given (user, item) pairs. The rough procedure for predicting the ratings is described as follows:

* We denote the target user and item as $u_m$ and $i_n$.
* We find a list of all neighboring users $N_k(u_m)$ of the target user, where distance is measured by cosine similarity.
* For each nearby user $u'$, we find all reviews $(u', i', r) \in R(u')$ made by this user.
* For each user, we save the tuple $(\text{sim}(u_m, u'), \text{sim}(i_n, i'), r)$ in a list `score_info`.
* We compute the predicted score using the `score_info` list by taking the weighted some of the scores $r$ by weights $\text{sim}(u_m, u') \cdot \text{sim}(i_n, i')$.

Concerely, we use the following formula for item score prediction: $$ \hat{r}(u_m, i_n) = \frac{\sum_{u' \in N_k(u_m)} \sum_{(u', i', r) \in R(u')} \text{sim}(u_m, u') \cdot \text{sim}(i_n, i') \cdot r}{\sum_{u' \in N_k(u_m)} \sum_{(u', i', r) \in R(u')} \text{sim}(u_m, u') \cdot \text{sim}(i_n, i')}.$$

[1] Hernández-Rubio, M., Cantador, I., Bellogín, A. 2019. A comparative analysis of recommender systems based on item aspect opinions extracted from user reviews. User Modeling and User-Adapted Interaction 29(2), 381-441.

def rate_item(user_vector, item_id, k=10):
    # fetch item vector
    item_vector = np.array(df_item[df_item["item_id"] == item_id].iloc[0]["aspect_weights"]).astype(float)

    # get the nearest k users given the user vector
    nearest_user_dists, nearest_users = index.search(user_vector.reshape((1,d)).astype(np.float32), k)
    nearest_user_dists = nearest_user_dists / np.linalg.norm(nearest_user_dists)

    score_info = []

    # for each nearby user...
    for user_ix in nearest_users[0]:
        # get his/her similarity value the queried user
        user_id = df_user.iloc[user_ix]["user_id"]
        nearby_vector = np.array(df_user[df_user["user_id"] == user_id].iloc[0]["aspect_weights"]).astype(float)
        user_sim = np.dot(user_vector, nearby_vector)/(np.linalg.norm(user_vector) * np.linalg.norm(nearby_vector))

        # iterate all his/her reviews
        reviews = df_review[df_review["user_id"] == user_id]
        for _, review in reviews.iterrows():
            # record the (user similarity, item similarity, reviewed score) tuple
            item = df_item[df_item["item_id"] == review["business_id"]]
            if len(item) == 0:
                continue
            review_vector = np.array(item.iloc[0]["aspect_weights"]).astype(float)
            item_sim = np.dot(item_vector, review_vector)/(np.linalg.norm(item_vector) * np.linalg.norm(review_vector))
            score_info.append([user_sim, item_sim, review["stars"]])

    # accumulate results
    weights = np.array([x[0] * x[1] for x in score_info])
    weights /= np.sum(weights)
    ratings = np.array([x[2] for x in score_info])
    return weights.dot(ratings)

rate_item(user_vector, "XOSRcvtaKc_Q5H1SAzN20A")

def predict_score(user, item_id):
    # sample user input: { 
    #     "latitude": 40,
    #     "longitude": -80,
    #     "...": (other info),
    #     "reviews": [
    #         {
    #             "item_id": "A",
    #             "rating": 4.5
    #         }
    #     ]
    # }
    return rate_item(get_user_vector(user["reviews"]), item_id)

## Process Reviews

Clean up the reviews dataframe and save to `reviews.json`.

df_item_to_join = df_item[['aspect_weights', 'item_id']].rename(columns={'aspect_weights': 'business_vec'}).set_index('item_id')
df_user_to_join = df_user[['aspect_weights', 'user_id']].rename(columns={'aspect_weights': 'user_vec'}).set_index('user_id')

df_review_join = df_review.drop(columns=['useful', 'funny', 'cool', 'date']).join(df_item_to_join, on='business_id').join(df_user_to_join, on='user_id').dropna()

user_avg = df_review_join.groupby(['user_id'])[['user_id', 'stars']].mean().rename(columns={'stars': 'user_avg_stars'})
df_review_join = df_review_join.join(user_avg, on='user_id')

df_review_join = df_review_join.reset_index().drop(columns=['index'])

drop_business_vec = np.array(df_review_join['business_vec'].tolist(), dtype=float)
drop_business_vec = np.sum(np.isnan(business_vec), axis=1).astype(int)
drop_user_vec = np.array(df_review_join['user_vec'].tolist(), dtype=float)
drop_user_vec = np.sum(np.isnan(user_vec), axis=1).astype(int)
drop_indices = np.logical_or(drop_business_vec, drop_user_vec)

df_review_join['drop_flag'] = pd.Series(drop_indices)

df_review_join = df_review_join[df_review_join['drop_flag'] == False].drop(columns=['drop_flag'])
df_review_join = df_review_join.reset_index().drop(columns=['index'])

df_review_join.tail()

df_review_join.to_json('../data/generated/review.json')
