==> notebooks/2_evaluation_metrics.ipynb <==
import os
import numpy as np
import pandas as pd
from tqdm import tqdm

from dataset import YelpReviewsDataset
from recommendation_engine import CFARecommendationEngine, RandomRecommendationEngine

dataset = YelpReviewsDataset()

## RMSE Error

Given a dataset split, the RMSE error calculates the root mean squared prediction error of a recommendation engine on a sample of user ratings in the dataset split.

def rmse(model, dataset, mode='val', n_samples=64):
    errors = []
    ids = dataset.get_ids(mode)
    if n_samples is not None:
        ids = np.random.choice(ids, n_samples, replace=False)
    for id in tqdm(ids):
        data_point = dataset[id]
        predicted = model.predict_score(data_point['user'], data_point['item_id'])
        gt = dataset[id]['stars']
        errors.append(predicted - gt)
    errors = np.array(errors)[~np.isnan(errors)]
    return np.sqrt(np.mean(np.square(errors)))

random_model_rmse = rmse(RandomRecommendationEngine(), dataset, mode='val')
print("RMSE of Random Model: {}".format(random_model_rmse))

random_model_rmse = rmse(CFARecommendationEngine(df_review=dataset.dataset_train), dataset, mode='val')
print("RMSE of Aspect-based Collaborative Filtering Model: {}".format(random_model_rmse))
